Rename BatchPropagationResults to MonteCarloResults and break out into own processor
