+++ adam/job.py
@@ -1,6 +1,8 @@
 """
     job.py
 """
+import urllib
+from adam import Project
 
 
 class Job(object):
@@ -22,7 +24,6 @@ class Job(object):
         self._execution_start = execution_start
         self._completion_time = completion_time
         self._status = status
-
 
     def __repr__(self):
         return (
@@ -66,3 +67,99 @@ class Job(object):
     def get_status(self):
         return self._status
 
+
+class Jobs(object):
+    """Module for managing jobs.
+
+    """
+
+    def __init__(self, rest):
+        """Initialize the Jobs API client.
+
+        Args:
+            rest (RestProxy): a RestProxy that makes calls to the ADAM API.
+        """
+        self._rest = rest
+
+    def find_jobs(self, project, status=None, object_id=None, user_defined_id=None, description=None,
+                  earliest_submission_datetime=None, latest_submission_datetime=None):
+        """Finds jobs based on one or more of the specified fields
+
+        Args:
+            project (str | Project): (Required) The workspace (project) id or a Project object
+            status (str): (Optional) The status field looking for (can have wildcards)
+            object_id (str): (Optional) The object ID field looking for (can have wildcards)
+            user_defined_id (str): (Optional) The user ID field looking for (can have wildcards)
+            description (str): (Optional) The description field looking for (can have wildcards)
+            earliest_submission_datetime (datetime): (Optional) Earlist submission date/time interested in
+            latest_submission_datetime (datetime): (Optional) Earlist submission date/time interested in
+
+        Returns:
+            list: a list of jobs for the project that match that criteria.
+        """
+
+        project_id = project.get_uuid() if type(project) is Project else project
+        query_parameters = []
+
+        if (status is not None):
+            query_parameters.append(f"status={urllib.parse.quote_plus(status)}")
+
+        if (object_id is not None):
+            query_parameters.append(f"objectId={urllib.parse.quote_plus(object_id)}")
+
+        if (user_defined_id is not None):
+            query_parameters.append(f"userDefinedId={urllib.parse.quote_plus(user_defined_id)}")
+
+        if (description is not None):
+            query_parameters.append(f"description={urllib.parse.quote_plus(description)}")
+
+        if (earliest_submission_datetime is not None):
+            if type(earliest_submission_datetime) is not datetime.datetime:
+                raise TypeError("Earliest submission date time needs to be a datetime object")
+            query_parameters.append(f"earliestSubmissionDateTime={earliest_submission_datetime.isoformat()}")
+
+        if (latest_submission_datetime is not None):
+            if type(latest_submission_datetime) is not datetime.datetime:
+                raise TypeError("Earliest submission date time needs to be a datetime object")
+            query_parameters.append(f"latestSubmissionDateTime={latest_submission_datetime.isoformat()}")
+
+        query_string = "&".join(query_parameters)
+        request_path = f'/projects/{project_id}/jobs?{query_string}'
+        project_id = project.get_uuid() if type(project) is Project else project
+        code, response = self._rest.get(request_path)
+
+        if (code == 200):
+            jobs = list(map(self._jobObjectFromHashMap, response['items']))
+            return jobs
+        else:
+            raise RuntimeError("Server status code: %s; Response: %s" % (code, response))
+
+    def _jobObjectFromHashMap(self, j):
+        try:
+            submission_time = dateparser.parse(j['submissionTime'])
+        except:
+            submission_time = None
+
+        try:
+            execution_start = dateparser.parse(j['executionStart'])
+        except:
+            execution_start = None
+
+        try:
+            completion_time = dateparser.parse(j['completionTime'])
+        except:
+            completion_time = None
+
+        return Job(
+            uuid=j.get('uuid'),
+            project_id=j.get('referenceUuid'),
+            description=j.get('description'),
+            object_id=j.get('objectId'),
+            user_defined_id=j.get('userDefinedId'),
+            job_type=j.get('jobType'),
+            input_json=j.get('inputParametersJson'),
+            submission_time=submission_time,
+            execution_start=execution_start,
+            completion_time=completion_time,
+            status=j.get('status')
+        )
